Temporarily adding components via the Sorter no longer triggers add/change events
Utils/Sorter will temporarily add a component to the collection in order
to capture its properties, then remove it immediately. This triggers the
component:add, component:remove, and change:changesCount events (and
possibly others).

The temporary addition/removal now includes an option flag 'temporary'
which can be used in event handlers to detect this scenario. In
addition, the editor.componentsUpdated handler performs the same check
which prevents updating the changesCount as well as updating storage.